Refactoring_Pastes_Structure
