+++ package.json
@@ -27,6 +27,7 @@
         "@storybook/manager-webpack5": "^6.5.16",
         "antd": "^5.19.4",
         "axios": "^0.27.2",
+        "dayjs": "^1.11.13",
         "next": "^12.1.0",
         "next-images": "^1.8.2",
         "nookies": "^2.5.2",
