+++ src/components/ImagePiker/styles.ts
