+++ src/interfaces/Address.ts
@@ -1,0 +1,11 @@
+export interface Address {
+    id: string;
+    street: string;
+    numberHouse: number;
+    neighborhood: string;
+    city: string;
+    state: string;
+    cep: string;
+    companyId: string;
+    userId: string | null;
+}
