+++ src/pages/Companies/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import CompaniesPage from 'template/CompaniesPage';
+
+const Companies: React.FC = () => {
+    return <CompaniesPage />;
+};
+
+export default Companies;
