+++ src/pages/Coupons/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import CouponsPage from 'template/CouponsPage';
+
+const Coupons: React.FC = () => {
+    return <CouponsPage />;
+};
+
+export default Coupons;
