+++ src/pages/CreateCompany/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import CreateCompanyPage from 'template/CreateCompanyPage';
+
+const CreateCompany: React.FC = () => {
+    return <CreateCompanyPage />;
+};
+
+export default CreateCompany;
