+++ src/pages/CreateCoupon/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import CreateCouponPage from 'template/CreateCouponPage';
+
+const CreateCoupon: React.FC = () => {
+    return <CreateCouponPage />;
+};
+
+export default CreateCoupon;
