+++ src/pages/CreatePost/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import CreatePostPage from 'template/CreatePostPage';
+
+const CreatePost: React.FC = () => {
+    return <CreatePostPage />;
+};
+
+export default CreatePost;
