+++ src/pages/Events/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import EventsPage from 'template/EventsPage';
+
+const Events: React.FC = () => {
+    return <EventsPage />;
+};
+
+export default Events;
