+++ src/pages/Plans/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import PlansPage from 'template/PlansPage';
+
+const Plans: React.FC = () => {
+    return <PlansPage />;
+};
+
+export default Plans;
