+++ src/pages/Posts/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import PostsPage from 'template/PostsPage';
+
+const Posts: React.FC = () => {
+    return <PostsPage />;
+};
+
+export default Posts;
