+++ src/pages/ShowUsers/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import ShowUsersPage from 'template/ShowUsersPage';
+
+const ShowUsers: React.FC = () => {
+    return <ShowUsersPage />;
+};
+
+export default ShowUsers;
