+++ src/pages/Users/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import UsersPage from 'template/UsersPage';
+
+const Users: React.FC = () => {
+    return <UsersPage />;
+};
+
+export default Users;
