+++ src/template/CompaniesPage/index.tsx
@@ -1,0 +1,82 @@
+/* eslint-disable no-console */
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import React, { useState, useEffect } from 'react';
+import { Table } from 'antd';
+import { PlusOutlined } from '@ant-design/icons';
+import { useRouter } from 'next/router';
+import { Company } from 'interfaces/Companies';
+import CompanyService from 'services/CompanyService';
+import { getColumns } from './index-helper';
+
+const CompaniesPage = () => {
+    const [companiesList, setCompaniesList] = useState<Company[]>([]);
+    const router = useRouter();
+
+    useEffect(() => {
+        const getCompanies = async () => {
+            try {
+                const response = await CompanyService.GetAll();
+                setCompaniesList(response);
+            } catch (error) {
+                console.error('Failed to fetch companies:', error);
+            }
+        };
+
+        getCompanies();
+    }, []);
+
+    const handleEdit = (company: Company) => {
+        // Navega para a página de cupons com os valores da empresa
+        router.push({
+            pathname: 'Coupons',
+            query: { initialValues: JSON.stringify(company) }
+        });
+    };
+
+    const handleDelete = async (id: string) => {
+        try {
+            await CompanyService.deleteCompany(id);
+            const updatedCompanies = await CompanyService.GetAll();
+            setCompaniesList(updatedCompanies);
+        } catch (error) {
+            console.error('Failed to delete company:', error);
+        }
+    };
+
+    const handleCreateCompany = () => {
+        // Navega para a página de criação de empresa
+        router.push('CreateCompany');
+    };
+
+    return (
+        <>
+            <div
+                style={{
+                    display: 'flex',
+                    justifyContent: 'space-between',
+                    alignItems: 'center',
+                    marginBottom: 8,
+                    marginTop: 8,
+                    marginRight: 16
+                }}
+            >
+                <h2>Empresas</h2>
+                <PlusOutlined
+                    style={{
+                        fontSize: '32px',
+                        color: '#CC8D3E'
+                    }}
+                    onClick={handleCreateCompany}
+                />
+            </div>
+            <Table
+                style={{ color: 'white' }}
+                columns={getColumns(handleEdit, handleDelete)}
+                dataSource={companiesList}
+                rowKey="id"
+            />
+        </>
+    );
+};
+
+export default CompaniesPage;
